Addressed changes in #60. Split data classes into old and new versions. Moved all non-constants out of constants and into conversion.py. Found source for codes.csv
